Updates DC list and allows unknown DC designators
Now, rather than raising an index error when a DC designator cannot be
found, we go ahead and try to use the provided DC name to construct the
auth URL anyway. This at least allows Swift in new DCs to be used
without it needing to exist in the DC dict. Folks won't necessarily
have to upgrade this package just to add new DCs.

This doesn't fully remove the object_storage.DATACENTERS dict since
someone could conceivably be referencing this list for something -- but
it will be deprecated as constructing on the fly removes the need for
these to be pre-defined.